Move creation of completion provider to ThreadPool
We do not block the caller if providers are not already available.
