+++ src/Features/Core/Portable/Completion/CompletionServiceWithProviders.cs
@@ -9,18 +9,13 @@ using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Diagnostics;
 using System.Linq;
-using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.Completion.Providers;
-using Microsoft.CodeAnalysis.Diagnostics;
-using Microsoft.CodeAnalysis.Host.Mef;
 using Microsoft.CodeAnalysis.LanguageServices;
 using Microsoft.CodeAnalysis.Options;
 using Microsoft.CodeAnalysis.PooledObjects;
 using Microsoft.CodeAnalysis.Shared.Extensions;
-using Microsoft.CodeAnalysis.Shared.Utilities;
 using Microsoft.CodeAnalysis.Text;
 using Roslyn.Utilities;
 
@@ -29,33 +24,15 @@ namespace Microsoft.CodeAnalysis.Completion
     /// <summary>
     /// A subtype of <see cref="CompletionService"/> that aggregates completions from one or more <see cref="CompletionProvider"/>s.
     /// </summary>
-    public abstract partial class CompletionServiceWithProviders : CompletionService, IEqualityComparer<ImmutableHashSet<string>>
+    public abstract partial class CompletionServiceWithProviders : CompletionService
     {
-        private readonly object _gate = new();
-
-        private readonly ConditionalWeakTable<IReadOnlyList<AnalyzerReference>, StrongBox<ImmutableArray<CompletionProvider>>> _projectCompletionProvidersMap
-             = new();
-
-        private readonly ConditionalWeakTable<AnalyzerReference, ProjectCompletionProvider> _analyzerReferenceToCompletionProvidersMap
-            = new();
-        private readonly ConditionalWeakTable<AnalyzerReference, ProjectCompletionProvider>.CreateValueCallback _createProjectCompletionProvidersProvider
-            = new(r => new ProjectCompletionProvider(r));
-
-        private readonly Dictionary<string, CompletionProvider> _nameToProvider = new();
-        private readonly Dictionary<ImmutableHashSet<string>, ImmutableArray<CompletionProvider>> _rolesToProviders;
-        private readonly Func<ImmutableHashSet<string>, ImmutableArray<CompletionProvider>> _createRoleProviders;
-        private readonly Func<string, CompletionProvider> _getProviderByName;
-
         private readonly Workspace _workspace;
-
-        private IEnumerable<Lazy<CompletionProvider, CompletionProviderMetadata>> _importedProviders;
+        private readonly ProviderSource _providerSource;
 
         protected CompletionServiceWithProviders(Workspace workspace)
         {
             _workspace = workspace;
-            _rolesToProviders = new Dictionary<ImmutableHashSet<string>, ImmutableArray<CompletionProvider>>(this);
-            _createRoleProviders = CreateRoleProviders;
-            _getProviderByName = GetProviderByName;
+            _providerSource = new ProviderSource(this);
         }
 
         public override CompletionRules GetRules()
@@ -69,59 +46,9 @@ namespace Microsoft.CodeAnalysis.Completion
         protected virtual ImmutableArray<CompletionProvider> GetBuiltInProviders()
             => ImmutableArray<CompletionProvider>.Empty;
 
-        private IEnumerable<Lazy<CompletionProvider, CompletionProviderMetadata>> GetImportedProviders()
-        {
-            if (_importedProviders == null)
-            {
-                var language = Language;
-                var mefExporter = (IMefHostExportProvider)_workspace.Services.HostServices;
-
-                var providers = ExtensionOrderer.Order(
-                        mefExporter.GetExports<CompletionProvider, CompletionProviderMetadata>()
-                        .Where(lz => lz.Metadata.Language == language)
-                        ).ToList();
-
-                Interlocked.CompareExchange(ref _importedProviders, providers, null);
-            }
-
-            return _importedProviders;
-        }
-
-        private ImmutableArray<CompletionProvider> CreateRoleProviders(ImmutableHashSet<string> roles)
-        {
-            var providers = GetAllProviders(roles);
-
-            foreach (var provider in providers)
-            {
-                _nameToProvider[provider.Name] = provider;
-            }
-
-            return providers;
-        }
-
-        private ImmutableArray<CompletionProvider> GetAllProviders(ImmutableHashSet<string> roles)
-        {
-            var imported = GetImportedProviders()
-                .Where(lz => lz.Metadata.Roles == null || lz.Metadata.Roles.Length == 0 || roles.Overlaps(lz.Metadata.Roles))
-                .Select(lz => lz.Value);
-
-#pragma warning disable 0618
-            // We need to keep supporting built-in providers for a while longer since this is a public API.
-            // https://github.com/dotnet/roslyn/issues/42367
-            var builtin = GetBuiltInProviders();
-#pragma warning restore 0618
-
-            return imported.Concat(builtin).ToImmutableArray();
-        }
-
         protected ImmutableArray<CompletionProvider> GetProviders(ImmutableHashSet<string> roles)
         {
-            roles ??= ImmutableHashSet<string>.Empty;
-
-            lock (_gate)
-            {
-                return _rolesToProviders.GetOrAdd(roles, _createRoleProviders);
-            }
+            return _providerSource.GetProviders(roles ?? ImmutableHashSet<string>.Empty);
         }
 
         private ConcatImmutableArray<CompletionProvider> GetFilteredProviders(
@@ -136,48 +63,6 @@ namespace Microsoft.CodeAnalysis.Completion
             ImmutableHashSet<string> roles, CompletionTrigger trigger)
         {
             return GetProviders(roles);
-        }
-
-        private ImmutableArray<CompletionProvider> GetProjectCompletionProviders(Project project)
-        {
-            if (project is null)
-            {
-                return ImmutableArray<CompletionProvider>.Empty;
-            }
-
-            if (project is null || project.Solution.Workspace.Kind == WorkspaceKind.Interactive)
-            {
-                // TODO (https://github.com/dotnet/roslyn/issues/4932): Don't restrict completions in Interactive
-                return ImmutableArray<CompletionProvider>.Empty;
-            }
-
-            if (_projectCompletionProvidersMap.TryGetValue(project.AnalyzerReferences, out var completionProviders))
-            {
-                return completionProviders.Value;
-            }
-
-            return GetProjectCompletionProvidersSlow(project);
-
-            // Local functions
-            ImmutableArray<CompletionProvider> GetProjectCompletionProvidersSlow(Project project)
-            {
-                return _projectCompletionProvidersMap.GetValue(project.AnalyzerReferences, pId => new StrongBox<ImmutableArray<CompletionProvider>>(ComputeProjectCompletionProviders(project))).Value;
-            }
-
-            ImmutableArray<CompletionProvider> ComputeProjectCompletionProviders(Project project)
-            {
-                using var _ = ArrayBuilder<CompletionProvider>.GetInstance(out var builder);
-                foreach (var reference in project.AnalyzerReferences)
-                {
-                    var projectCompletionProvider = _analyzerReferenceToCompletionProvidersMap.GetValue(reference, _createProjectCompletionProvidersProvider);
-                    foreach (var completionProvider in projectCompletionProvider.GetExtensions(project.Language))
-                    {
-                        builder.Add(completionProvider);
-                    }
-                }
-
-                return builder.ToImmutable();
-            }
         }
 
         private ImmutableArray<CompletionProvider> FilterProviders(
@@ -228,19 +113,10 @@ namespace Microsoft.CodeAnalysis.Completion
 
             if (item.ProviderName != null)
             {
-                lock (_gate)
-                {
-                    provider = _nameToProvider.GetOrAdd(item.ProviderName, _getProviderByName);
-                }
+                _providerSource.GetProviderByName(item.ProviderName);
             }
 
             return provider;
-        }
-
-        private CompletionProvider GetProviderByName(string providerName)
-        {
-            var providers = GetAllProviders(roles: ImmutableHashSet<string>.Empty);
-            return providers.FirstOrDefault(p => p.Name == providerName);
         }
 
         public override async Task<CompletionList> GetCompletionsAsync(
@@ -576,40 +452,6 @@ namespace Microsoft.CodeAnalysis.Completion
             }
         }
 
-        bool IEqualityComparer<ImmutableHashSet<string>>.Equals(ImmutableHashSet<string> x, ImmutableHashSet<string> y)
-        {
-            if (x == y)
-            {
-                return true;
-            }
-
-            if (x.Count != y.Count)
-            {
-                return false;
-            }
-
-            foreach (var v in x)
-            {
-                if (!y.Contains(v))
-                {
-                    return false;
-                }
-            }
-
-            return true;
-        }
-
-        int IEqualityComparer<ImmutableHashSet<string>>.GetHashCode(ImmutableHashSet<string> obj)
-        {
-            var hash = 0;
-            foreach (var o in obj)
-            {
-                hash += o.GetHashCode();
-            }
-
-            return hash;
-        }
-
         private class DisplayNameToItemsMap : IEnumerable<CompletionItem>, IDisposable
         {
             // We might need to handle large amount of items with import completion enabled,
@@ -715,7 +557,10 @@ namespace Microsoft.CodeAnalysis.Completion
                 => _completionServiceWithProviders = completionServiceWithProviders;
 
             internal ImmutableArray<CompletionProvider> GetAllProviders(ImmutableHashSet<string> roles)
-                => _completionServiceWithProviders.GetAllProviders(roles);
+                => _completionServiceWithProviders._providerSource.GetProviders(roles, waitUntilAvaialble: true);
+
+            internal void EnensureCompeltionProvidersAvailable()
+                => _completionServiceWithProviders._providerSource.GetProviders(ImmutableHashSet<string>.Empty, waitUntilAvaialble: true);
 
             internal Task<CompletionContext> GetContextAsync(
                 CompletionProvider provider,
